Refactorizado de board + tests funcionando
